+++ src/compiler/sema_decls.c
@@ -3306,6 +3306,24 @@ INLINE bool sema_analyse_attribute_bool_const(SemaContext *context, Expr *expr)
 	return true;
 }
 
+INLINE StackProtector stack_protector_from_string(const char *name)
+{
+	if (str_eq(name, "none")) return STACK_PROTECTOR_NONE;
+	if (str_eq(name, "basic")) return STACK_PROTECTOR_BASIC;
+	if (str_eq(name, "strong")) return STACK_PROTECTOR_STRONG;
+	if (str_eq(name, "all")) return STACK_PROTECTOR_ALL;
+	return STACK_PROTECTOR_NOT_SET;
+}
+
+INLINE StackProbe stack_probe_from_string(const char *name)
+{
+	if (str_eq(name, "none")) return STACK_PROBE_NONE;
+	if (str_eq(name, "call")) return STACK_PROBE_CALL;
+	if (str_eq(name, "inline")) return STACK_PROBE_INLINE;
+	return STACK_PROBE_NOT_SET;
+}
+
+
 /**
  * Analyse almost all attributes.
  */
@@ -3321,6 +3339,10 @@ static bool sema_analyse_attribute(SemaContext *context, ResolvedAttrData *attr_
 			[ATTRIBUTE_BIGENDIAN] = ATTR_BITSTRUCT,
 			[ATTRIBUTE_BUILTIN] = ATTR_MACRO | ATTR_FUNC | ATTR_GLOBAL | ATTR_CONST,
 			[ATTRIBUTE_CALLCONV] = ATTR_FUNC | ATTR_INTERFACE_METHOD | ATTR_FNTYPE,
+			[ATTRIBUTE_STACKPROBE] = ATTR_FUNC | ATTR_FNTYPE,
+			[ATTRIBUTE_NOSTACKPROBE] = ATTR_FUNC | ATTR_FNTYPE,
+			[ATTRIBUTE_STACKPROTECTOR] = ATTR_FUNC | ATTR_FNTYPE,
+			[ATTRIBUTE_NOSTACKPROTECTOR] = ATTR_FUNC | ATTR_FNTYPE,
 			[ATTRIBUTE_CNAME] = ATTR_FUNC | ATTR_GLOBAL | ATTR_CONST | USER_DEFINED_TYPES,
 			[ATTRIBUTE_COMPACT] = ATTR_STRUCT | ATTR_UNION,
 			[ATTRIBUTE_CONST] = ATTR_MACRO,
@@ -3425,6 +3447,28 @@ static bool sema_analyse_attribute(SemaContext *context, ResolvedAttrData *attr_
 			if (!expr) RETURN_SEMA_ERROR(decl, "Expected a string argument.");
 			if (!sema_analyse_attribute_string_const(context, expr)) return false;
 			if (!update_call_abi_from_string(context, decl, expr)) return false;
+			return true;
+		case ATTRIBUTE_STACKPROBE:
+			if (!expr) RETURN_SEMA_ERROR(attr, "Expected a string argument.");
+			if (!sema_analyse_attribute_string_const(context, expr)) return false;
+			if ((decl->func_decl.stack_probe = 1 + stack_probe_from_string(expr->const_expr.bytes.ptr)) == STACK_PROBE_NOT_SET + 1) // NOLINT
+			{
+				RETURN_SEMA_ERROR(expr, "Unknown stack probe level, only 'none', 'call' and 'inline' are supported");
+			}
+			return true;
+		case ATTRIBUTE_NOSTACKPROBE:
+			decl->func_decl.stack_probe = STACK_PROBE_NONE + 1;
+			return true;
+		case ATTRIBUTE_STACKPROTECTOR:
+			if (!expr) RETURN_SEMA_ERROR(attr, "Expected a string argument.");
+			if (!sema_analyse_attribute_string_const(context, expr)) return false;
+			if ((decl->func_decl.stack_protector = 1+ stack_protector_from_string(expr->const_expr.bytes.ptr)) == STACK_PROTECTOR_NOT_SET + 1) // NOLINT
+			{
+				RETURN_SEMA_ERROR(expr, "Unknown stack protection level, only 'none', 'basic', 'strong' and 'all' are supported");
+			}
+			return true;
+		case ATTRIBUTE_NOSTACKPROTECTOR:
+			decl->func_decl.stack_protector = STACK_PROTECTOR_NONE + 1;
 			return true;
 		case ATTRIBUTE_BENCHMARK:
 			decl->func_decl.attr_benchmark = true;